feat: remove `reqwest` feature
When the `http-request-derive` crate is used as a dependency twice, once
with and once without the `reqwest` feature, the feature will be enabled
automatically. That requires the gated trait methods to be implemented,
but they are not in the crate that requires `http-request-derive`
without the `reqwest` feature, resulting in a compile error. We remove
the feature for now, the required functionality will be introduced in a
different structure.